chore: move all configurables to configurables
