+++ lib/concerns/wallaby/authorizable.rb
@@ -3,56 +3,13 @@
 module Wallaby
   # Authorizer related attributes
   module Authorizable
-    extend ActiveSupport::Concern
-
-    # Configurable attribute for authorizer related
-    module ClassMethods
-      # @!attribute [w] model_authorizer
-      def model_authorizer=(model_authorizer)
-        ModuleUtils.inheritance_check model_authorizer, application_authorizer
-        @model_authorizer = model_authorizer
-      end
-
-      # @!attribute [r] model_authorizer
-      # If Wallaby doesn't get it right, please specify the **model_authorizer**.
-      # @example To set model authorizer
-      #   class Admin::ProductionsController < Admin::ApplicationController
-      #     self.model_authorizer = ProductAuthorizer
-      #   end
-      # @return [Class] model authorizer
-      # @raise [ArgumentError] when **model_authorizer** doesn't inherit from **application_authorizer**
-      # @see Wallaby::ModelAuthorizer
-      # @since wallaby-5.2.0
-      attr_reader :model_authorizer
-
-      # @!attribute [w] application_authorizer
-      def application_authorizer=(application_authorizer)
-        ModuleUtils.inheritance_check model_authorizer, application_authorizer
-        @application_authorizer = application_authorizer
-      end
-
-      # @!attribute [r] application_authorizer
-      # The **application_authorizer** is as the base class of {#model_authorizer}.
-      # @example To set application decorator:
-      #   class Admin::ApplicationController < Wallaby::ResourcesController
-      #     self.application_authorizer = AnotherApplicationAuthorizer
-      #   end
-      # @return [Class] application decorator
-      # @raise [ArgumentError] when **model_authorizer** doesn't inherit from **application_authorizer**
-      # @see Wallaby::ModelAuthorizer
-      # @since wallaby-5.2.0
-      def application_authorizer
-        @application_authorizer || superclass.try(:application_authorizer)
-      end
-    end
-
     # Model authorizer for current modal class.
     #
     #  It can be configured in following class attributes:
     #
-    # - controller configuration {Wallaby::Authorizable::ClassMethods#model_authorizer .model_authorizer}
+    # - controller configuration {Wallaby::Configurable::ClassMethods#model_authorizer .model_authorizer}
     # - a generic authorizer based on
-    #   {Wallaby::Authorizable::ClassMethods#application_authorizer .application_authorizer}
+    #   {Wallaby::Configurable::ClassMethods#application_authorizer .application_authorizer}
     # @return [Wallaby::ModelAuthorizer] model authorizer
     # @since wallaby-5.2.0
     def current_authorizer
