+++ lib/concerns/wallaby/engineable.rb
@@ -3,28 +3,6 @@
 module Wallaby
   # Engine related helper methods for both controller and view
   module Engineable
-    extend ActiveSupport::Concern
-
-    # Configurable attribute
-    module ClassMethods
-      # @!attribute [w] engine_name
-      attr_writer :engine_name
-
-      # @!attribute [r] engine_name
-      # The engine name will be used to handle URLs.
-      #
-      # So when to set this engine name? When Wallaby doesn't know what is the correct engine helper to use.
-      # @example To set an engine name:
-      #   class Admin::ApplicationController < Wallaby::ResourcesController
-      #     self.engine_name = 'admin_engine'
-      #   end
-      # @return [String, Symbol, nil] engine name
-      # @since wallaby-5.2.0
-      def engine_name
-        @engine_name || superclass.try(:engine_name)
-      end
-    end
-
     # This engine helper is used to access URL helpers of Wallaby engine.
     #
     # Considering **Wallaby** is mounted at the following paths:
